+++ src/main/php/text/ical/Role.class.php
@@ -1,0 +1,7 @@
+<?php namespace text\ical;
+
+abstract class Role {
+  const CHAIR           = 'CHAIR';
+  const REQ_PARTICIPANT = 'REQ-PARTICIPANT';
+  const NON_PARTICIPANT = 'NON-PARTICIPANT';
+}