+++ ChangeLog.md
@@ -3,6 +3,12 @@ ICal ChangeLog
 
 ## ?.?.? / ????-??-??
 
+* Introduced class constants:
+  - `text.ical.Method` for calendar methods, e.g. *REQUEST* or *REPLY*
+  - `text.ical.Role` for roles, e.g. *CHAIR* or *REQ-PARTICIPANT*.
+  - `text.ical.PartStat` for status, e.g. *NEEDS-ACTION* or *TENTATIVE*
+  (@thekid)
+
 ## 0.4.0 / 2016-06-26
 
 * Wrapped lines after 75 characters. The implementation will keep pairs